Set dynamic height to campaign browser cell ImageView width is set as per provided screenshot (width of the screen) and height with aspect ration of 4:3 Text is set to below image There is no space between image and title, Title can max be of two lines, Padding is set as required Description text has 8 points space on top and no height restriction.
In provided screenshot, image has 8 points padding on both left and right. And description statement is saying image should use full available width of the screen. So currently I have set 8 points leading and trailing to image in cell in storyboard as per screenshot. If we need to follow description we can update leading and trailing of imageView from 8 to 0.